Updated the costume collections feature Additions: Ability to see a count of costumes in a collection (and owned count) Ability to filter, search, and sort collections (tools bar will automatically be shown when collection length is greater than 20). Added a share button (that copies a shorter link - id is not included, since it's a per browser property) Added a New Collection button to the Collections List. Added a confirmation step when cancelling the creation/editing of a collection (if there are any selected items) Added an official collection for the 616 Vault (2026)
Fixed a reported issue with Starlord's Groovy Guardian nameplate having the wrong image and the Glory Gunner nameplate missing.

Improved automation scripts for costumes, nameplates, and achievements. The first two were modified to only add new items based on a diff, rather than always overwriting the item lists.

+++ scripts/parse-achievements-table/index.ts
@@ -171,14 +171,16 @@ function parseAchievementCategory(category: AchievementTypeCategory) {
 	const diff = createDiff(achievements, previousAchivements);
 
 	if (diff.length) {
-		p.log.info(`Making a backup of the previous achievements file (${category})...`);
 		// make a backup in case things go south
-		fs.copyFileSync(
-			outputAppPath,
-			ACHIEVEMENTS_FILE_BACKUP(category).replace(
-				'%DATE%', fileNameFriendlyDate(new Date())
-			)
-		);
+		if (fs.existsSync(outputAppPath)) {
+			p.log.info(`Making a backup of the previous achievements file (${category})...`);
+			fs.copyFileSync(
+				outputAppPath,
+				ACHIEVEMENTS_FILE_BACKUP(category).replace(
+					'%DATE%', fileNameFriendlyDate(new Date())
+				)
+			);
+		}
 
 		const combinedAchievements: (AchievementType|'_____new_achievements_____')[] = previousAchivements.slice();
 		combinedAchievements.push(
@@ -190,6 +192,7 @@ function parseAchievementCategory(category: AchievementTypeCategory) {
 		fs.writeFileSync(outputAppPath, JSON.stringify(combinedAchievements, null, 2));
 
 		p.log.info(`Parsed ${category} successfully. Saved at: ${outputPath}, ${outputAppPath}`);
+		p.log.info(`Added ${diff.length} achievements to ${category}.`);
 	}
 	else
 		p.log.warn(`${category} didn't have any new additions. Nothing was modified.`);
